requirements.txt
chart_studio==1.1.0
cufflinks==0.17.3
mysql-connector-python==8.0.33
requests==2.29.0
numpy==1.24.2
pandas==1.5.3
plotly==5.13.1
seaborn==0.12.2